rough chop of test files, probably not building
