+++ state/lease/client_race_test.go
@@ -1,0 +1,25 @@
+// Copyright 2015 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package lease_test
+
+import (
+	"time"
+
+	jc "github.com/juju/testing/checkers"
+	gc "gopkg.in/check.v1"
+	"gopkg.in/mgo.v2/bson"
+
+	"github.com/juju/juju/state/lease"
+)
+
+// ClientRaceSuite tests the ugliest of details.
+type ClientRaceSuite struct {
+	FixtureSuite
+}
+
+var _ = gc.Suite(&ClientRaceSuite{})
+
+func (s *ClientRaceSuite) TestMany(c *gc.C) {
+	c.Fatalf("not done")
+}
